Storybook: Dropdown 스토리 작성
